[Fetcher] Moves get_args to a separate file, args.py
